backend/requirements.txt
python-dotenv==1.0.0
google-generativeai
firebase-admin
flask
werkzeug